NodeBox runs on 10.5-10.7 using Python 2.5.
We link in python2.5.dylib instead of the Mac Python framework (which
links to 2.7 on Lion). This makes NodeBox run on 10.5 - 10.7.  Howerver,
native extensions are still borked, erring out with "unknown required
load command 0x80000022". This is not a problem for cGeo and
cPathmatics, which are just faster versions of built-in functions, but
cPolymagic doesn't have a Python implementation.