banished inheritFromContext
- state-mixins now use properties to to merge customized and
context-derived values
- inheritance now happens at drawtime rather than as a separate step.
in addition it now defaults to drawing a copy. this breaks the behavior
where you can configure a primitive after it’s already been added to
the context, but that heisenbergian state is actually a minefield.
- the old behavior can still be achieved by passing copy=False to the
grob’s .draw() method.
- might make sense to have this be a plotstyle() mode. COPY/LIVE/OFF?
- Text & Stylesheet are mostly updated, but could use some tightening
up when the CT rewrite happens

[sidenote]
it looks like apple broke c-extensions with xcode 5.1. the
-mno-fused-madd in the interpreter’s CFLAGS causes clang to throw an
error now.

sadly the easiest fix until apple notices it goofed is to remove the
flag from
/System/Library/Frameworks/Python.framework/Versions/2.7/lib/python2.7/_
sysconfigdata.py and zap its pyc & pyo